Large performance increase. Braking works.
- waypoint_updater.py
Rate reduced from 40 to 0.1hz for a large performance increase
Don't really see the need to continously send the same base waypoints at all?!

- waypoint_updater.py
Refactored to be able to control update rate. So it is *not* only invoked when
new base_waypoints are sent (which now only happens every 10 seconds)

- dbw_node.py
Changed update rate from 50 to 10hz for performance
Added multiplier to brake value. Value range is much larger than for throttle
and doesn't react until many thousand units. Sending 20,000 to the brake_cmd
results in a reasonably strong brake input. This will have to be tweaked once
we actually stop the car.

- twist_controller.py
Set all contollers integral multipliers to 0.0 for now. Resulted in
compounding error and not important to have finely tuned and well-bounded
values for now.